FIL-71 pr fixes
