+++ core/vm/actor/actor.hpp
@@ -3,14 +3,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#ifndef CPP_FILECOIN_CORE_ACTOR_ACTOR_HPP
-#define CPP_FILECOIN_CORE_ACTOR_ACTOR_HPP
+#ifndef CPP_FILECOIN_CORE_VM_ACTOR_ACTOR_HPP
+#define CPP_FILECOIN_CORE_VM_ACTOR_ACTOR_HPP
 
 #include <libp2p/multi/content_identifier_codec.hpp>
 
 #include "primitives/big_int.hpp"
 
-namespace fc::actor {
+namespace fc::vm::actor {
   using primitives::BigInt;
   using libp2p::multi::ContentIdentifier;
 
@@ -43,6 +43,6 @@ namespace fc::actor {
   extern ContentIdentifier kAccountCodeCid, kCronCodeCid, kStoragePowerCodeCid,
       kStorageMarketCodeCid, kStorageMinerCodeCid, kMultisigCodeCid,
       kInitCodeCid, kPaymentChannelCodeCid;
-}  // namespace fc::actor
+}  // namespace fc::vm::actor
 
-#endif  // CPP_FILECOIN_CORE_ACTOR_ACTOR_HPP
+#endif  // CPP_FILECOIN_CORE_VM_ACTOR_ACTOR_HPP
