+++ test/core/CMakeLists.txt
@@ -3,10 +3,10 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-add_subdirectory(actor)
 add_subdirectory(clock)
 add_subdirectory(common)
 add_subdirectory(crypto)
 add_subdirectory(primitives)
 add_subdirectory(storage)
 add_subdirectory(codec)
+add_subdirectory(vm)
