+++ test/core/vm/CMakeLists.txt
@@ -1,0 +1,6 @@
+#
+# Copyright Soramitsu Co., Ltd. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+
+add_subdirectory(actor)
